Eliminate the state from FileInput
Simplify FileInput by moving the filename and error state tracking from
FileInput up to FileOrTextInput, because the other use of FileInput
(inside ShareInput) explicitly omits the filename rendering anyway.

This lets us remove the "hack" described in componentWillReceiveProps,
which was intended to propagate the 'clearData' operation to the
FileInput's state. Now that the state is in FileOrTextInput, it can be
updated directly.

This is the first in a series of changes inteded to generalize and
refactor FileInput in an effort to support selecting multiple share
files at once. It doesn't make sense for FileInput to track the state of
a list of files, since it's not designed to render multiple filenames.